requirements.txt
Pillow
RPi.GPIO
gpiozero
spidev
datetime
requests
python-dotenv
ratelimit
lgpio